# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cycle/current-phase": {
      "get": {
        "operationId": "CycleController_getCurrentPhase",
        "summary": "Récupérer la phase actuelle du cycle",
        "description": "Retourne la phase actuelle du cycle menstruel de l'utilisatrice connectée",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Phase actuelle retournée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentPhaseResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucun cycle trouvé ou suivi des cycles désactivé"
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cycle/config": {
      "get": {
        "operationId": "CycleController_getCycleConfig",
        "summary": "Récupérer la configuration du cycle",
        "description": "Retourne les paramètres personnalisés du cycle menstruel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuration récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleConfigResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Configuration non trouvée"
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "CycleController_updateCycleConfig",
        "summary": "Mettre à jour la configuration du cycle",
        "description": "Modifie les paramètres personnalisés du cycle menstruel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCycleConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration mise à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleConfigResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "Configuration créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleConfigResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cycle/periods": {
      "post": {
        "operationId": "CycleController_logPeriod",
        "summary": "Enregistrer des règles",
        "description": "Logger manuellement le début et la fin des règles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Règles enregistrées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogPeriodResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CycleController_getPeriodsHistory",
        "summary": "Récupérer l'historique des règles",
        "description": "Liste chronologique des règles avec statistiques",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre de cycles à retourner",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Nombre de cycles à ignorer",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "Date de début (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "Date de fin (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historique récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodsHistoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cycle/predictions": {
      "get": {
        "operationId": "CycleController_getCyclePredictions",
        "summary": "Obtenir les prédictions du cycle",
        "description": "Calculs intelligents des prochaines règles et ovulation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Prédictions calculées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CyclePredictionsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Pas assez de données pour les prédictions"
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cycle/calendar": {
      "get": {
        "operationId": "CycleController_getCycleCalendar",
        "summary": "Obtenir le calendrier du cycle",
        "description": "Vue calendrier avec phases, prédictions et événements",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Date de début (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "months",
            "required": false,
            "in": "query",
            "description": "Nombre de mois (défaut: 3)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calendrier généré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleCalendarResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Cycle"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "FitMe API",
    "description": "Documentation de l'API FitMe - Application de fitness basée sur le cycle menstruel",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CurrentPhaseResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "Indique si la requête a été traitée avec succès"
          },
          "data": {
            "description": "Données détaillées sur la phase actuelle du cycle",
            "allOf": [
              {
                "$ref": "#/components/schemas/CurrentPhaseDataDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "Vous êtes actuellement en Phase folliculaire - Énergie croissante, idéale pour commencer de nouveaux défis",
            "description": "Message descriptif sur la phase actuelle"
          }
        },
        "required": [
          "success",
          "data",
          "message"
        ]
      },
...
      "CycleConfigResponseDto": {
        "type": "object",
        "properties": {
          "config": {
            "description": "Configuration du cycle",
            "allOf": [
              {
                "$ref": "#/components/schemas/CycleConfigDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "Message de confirmation",
            "example": "Configuration du cycle récupérée avec succès"
          }
        },
        "required": [
          "config",
          "message"
        ]
      },
      "UpdateCycleConfigDto": {
        "type": "object",
        "properties": {
          "isCycleTrackingEnabled": {
            "type": "boolean",
            "description": "Activer/désactiver le suivi du cycle",
            "example": true
          },
          "usesExternalProvider": {
            "type": "boolean",
            "description": "Activer/désactiver le fournisseur externe",
            "example": false
          },
          "useMenopauseMode": {
            "type": "boolean",
            "description": "Activer/désactiver le mode ménopause",
            "example": false
          },
          "averageCycleLength": {
            "type": "number",
            "description": "Durée moyenne du cycle (21-40 jours)",
            "example": 28,
            "minimum": 21,
            "maximum": 40
          },
          "averagePeriodLength": {
            "type": "number",
            "description": "Durée moyenne des règles (3-10 jours)",
            "example": 5,
            "minimum": 3,
            "maximum": 10
          },
          "prefersManualInput": {
            "type": "boolean",
            "description": "Préférer la saisie manuelle",
            "example": false
          }
        }
      },
      "LogPeriodDto": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "description": "Date de début des règles",
            "example": "2024-01-15",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "description": "Date de fin des règles",
            "example": "2024-01-20",
            "format": "date"
          },
          "flowIntensity": {
            "type": "number",
            "description": "Intensité du flux (1=léger, 5=très fort)",
            "example": 3,
            "minimum": 1,
            "maximum": 5
          },
          "notes": {
            "type": "string",
            "description": "Notes additionnelles",
            "example": "Flux normal, quelques crampes le premier jour"
          }
        },
        "required": [
          "startDate"
        ]
      },
...
      "LogPeriodResponseDto": {
        "type": "object",
        "properties": {
          "period": {
            "description": "Cycle créé ou mis à jour",
            "allOf": [
              {
                "$ref": "#/components/schemas/PeriodDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "Message de confirmation",
            "example": "Règles enregistrées avec succès"
          },
          "isNewCycle": {
            "type": "boolean",
            "description": "Indique si un nouveau cycle a été créé",
            "example": true
          }
        },
        "required": [
          "period",
          "message",
          "isNewCycle"
        ]
      },
      "PeriodsHistoryResponseDto": {
        "type": "object",
        "properties": {
          "periods": {
            "description": "Liste des cycles/règles",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Nombre total de cycles",
            "example": 12
          },
          "averageCycleLength": {
            "type": "number",
            "description": "Durée moyenne du cycle",
            "example": 28.5
          },
          "averagePeriodLength": {
            "type": "number",
            "description": "Durée moyenne des règles",
            "example": 5.2
          },
          "regularityPercentage": {
            "type": "number",
            "description": "Pourcentage de cycles réguliers",
            "example": 85.5
          }
        },
        "required": [
          "periods",
          "total",
          "averageCycleLength",
          "averagePeriodLength",
          "regularityPercentage"
        ]
      },
...
    }
  }
}